Rename 'applications' to 'applicationList'
