+++ src/features/applicationList/applicationsPage.module.scss
